Split config schema validators into submodule
